+++ pkg/detector/library/driver.go
@@ -51,7 +51,7 @@ func NewDriver(libType string) (Driver, error) {
 	case ftypes.Pipenv, ftypes.Poetry, ftypes.Pip, ftypes.PythonPkg:
 		ecosystem = vulnerability.Pip
 		comparer = pep440.Comparer{}
-	case ftypes.PubSpec:
+	case ftypes.Pub:
 		ecosystem = vulnerability.Pub
 		comparer = compare.GenericComparer{}
 	case ftypes.Hex:
