+++ pkg/fanal/analyzer/language/dart/pub/testdata/empty.lock
