+++ pkg/fanal/analyzer/language/dart/pub/testdata/happy.lock
